Added Stay Awake files
The .bat file will run the .py file in a minimized window, and will also keep track of how many times it's moved for you. You can set different parameters at the top of the .py file that work the best for you, and then execute the .bat file through Windows Task Scheduler.